django_drf/requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
autopep8==1.4.4
Babel==2.7.0
backports.csv==1.0.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-bower==5.2.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-import-export==1.2.0
django-nose==1.4.6
django-seed==0.1.9
django-tables2==2.0.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==1.0.7
flake8==3.7.7
Flask==1.0.3
flask-jsontools==0.1.1.post0
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
idna==2.8
inflect==2.1.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
mccabe==0.6.1
nose==1.3.7
ntlm-auth==1.3.0
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
passlib==1.7.1
Pillow==6.0.0
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlacodegen==2.0.1
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.34.0
sqlparse==0.3.0
tablib==0.13.0
text-unidecode==1.2
tqdm==4.32.1
urllib3==1.25.3
Werkzeug==0.15.4
xlrd==1.2.0
xlwt==1.3.0
